shorts_factory/requirements.txt
moviepy
edge-tts
asyncio
elevenlabs
requests
python-dotenv
